refactor: add LoginModule (eagerly loaded at launch)
