+++ src/app/login/index.ts
@@ -1,0 +1,2 @@
+export * from './login.module';
+export * from './login.routes';
