+++ src/app/login/login.routes.ts
@@ -1,0 +1,10 @@
+import { Routes } from '@angular/router';
+
+import { LoginComponent } from './login/login.component';
+
+export const loginRoutes: Routes = [
+  {
+    path: '',
+    component: LoginComponent
+  }
+];
